+++ src/components/parameter-value-input.vue
@@ -1,5 +1,18 @@
 <template>
   <div>
+    <geojson-input
+      v-if="isGeojsonType"
+      :param-name="paramName"
+      :param-type="paramType"
+      :default-value="defaultValue"
+      :readonly="readonly"
+
+      :value="value"
+      @change="val => { $emit('change', val) }"
+      @input="val => { $emit('input', val) }"
+    />
+
+    <!-- TODO the following with v-else, but for now also showing the textarea -->
     <q-input
       class="round-borders q-pa-xs bg-green-1"
       :readonly="readonly"
@@ -17,8 +30,14 @@
 </template>
 
 <script>
+  import geojsonInput from 'components/geojson-input'
+
   export default {
     props: {
+      paramName: {
+        type: String,
+        required: true
+      },
       paramType: {
         type: String,
         required: true
@@ -38,6 +57,10 @@
       },
     },
 
+    components: {
+      geojsonInput,
+    },
+
     data() {
       return {
         paramValue: this.value,
@@ -45,6 +68,29 @@
     },
 
     computed: {
+      isGeojsonType() {
+        switch (this.paramType.toLowerCase()) {
+          case 'point':
+          case 'multipoint':
+          case 'linestring':
+          case 'multilinestring':
+          case 'polygon':
+          case 'multipolygon':
+          case 'geometrycollection':
+          case 'geometry':
+          // https://tools.ietf.org/html/rfc7946#section-3
+          case 'geojson':
+            return true
+
+          // https://tools.ietf.org/html/rfc7946#section-3.2
+          // case 'feature':
+          // case 'featurecollection': return ??
+
+          default:
+            return false
+        }
+      },
+
       inputProps() {
         let type = "text"
         let style = "font-family:monospace"
@@ -53,10 +99,7 @@
         switch (this.paramType.toLowerCase()) {
           case 'float':
           case 'int':
-            break
-
           case 'boolean':
-            style += ";width:3em"
             break
 
           case 'string':